Visitor pattern project
This project implements visitor pattern and other design principles.
This project uses a tree as an element and two visitors one to populate
tree and one to count total number of words and total number of unique
words. This project has been tested with input files that contains 10
million, 40 million and 100 million words. A sample input file with 1
thousand and 1 million words are attached for reference.

Refer to README.txt for more details.